Created the first rendition of the 8BallGame
Return Values and return Statements lesson from Automate the Boring Stuff with Python Programming 
Chapter 3 – Functions
expression is used with a return statement, the return value is what this expression evaluates to. For example, the following program defines a function that returns a different string depending on what number it is passed as an argument. Enter this code into the file editor and save it as 8BallGame.py: